Timing Fixes
Apparently, loading the config file takes 2 seconds which is seriously throwing off the detection loop on the first run. Added a timer to account for this time and add it in to the scan interval.

Also, as the script now loops internally there is no reason for the job to be restarted every minute. The default job duration is now 600 seconds.